pyproject.toml
[project]
name = "resno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jax[cuda12]>=0.6.0",
    "equinox>=0.12.2",
    "pandas>=2.2.3",
    "tqdm>=4.67.1",
    "optax>=0.2.4",
    "quicksig>=0.1.0",
    "nvidia-dali-cuda120>=1.50.0",
    "matplotlib>=3.10.3",
    "diffrax>=0.7.0",
]

[tool.uv.sources]
quicksig = { git = "https://github.com/luke-a-thompson/QuickSig" }
